+++ backend/app/api/routes/items.py
@@ -3,7 +3,8 @@ from typing import Any
 from fastapi import APIRouter, HTTPException
 from sqlmodel import func, select
 
-from app.api.deps import CurrentUser, SessionDep
+from app.api.deps.users import CurrentUser
+from app.api.deps.utils import SessionDep
 from app.model.users import Item, ItemCreate, ItemOut, ItemsOut, ItemUpdate, Message
 
 router = APIRouter()
