+++ backend/app/api/routes/login.py
@@ -5,7 +5,8 @@ from fastapi import APIRouter, Depends, HTTPException
 from fastapi.responses import HTMLResponse
 from fastapi.security import OAuth2PasswordRequestForm
 
-from app.api.deps import CurrentUser, SessionDep, get_current_active_superuser
+from app.api.deps.users import CurrentUser, get_current_active_superuser
+from app.api.deps.utils import SessionDep
 from app.core import security
 from app.core.config import settings
 from app.core.security import get_password_hash
