+++ backend/app/api/routes/users.py
@@ -3,11 +3,11 @@ from typing import Any
 from fastapi import APIRouter, Depends, HTTPException
 from sqlmodel import col, delete, func, select
 
-from app.api.deps import (
+from app.api.deps.users import (
     CurrentUser,
-    SessionDep,
     get_current_active_superuser,
 )
+from app.api.deps.utils import SessionDep
 from app.core.config import settings
 from app.core.security import get_password_hash, verify_password
 from app.crud import users
